(Refatoração)[advogado_views, services_advogados] Concluida a refatoração do dominio advogado #2 Corrigindo duplicidades
